requirements.txt
fastapi
uvicorn
numpy
pandas
pyarrow
rasterio
pyproj
shapely
pydantic
faiss-cpu
tqdm
